The interfaces and ts types were moved to the respective folder and files to improve the project structure.
