--- src/types/forms.ts
@@ -1,13 +1,0 @@
-export interface DigitInputProps {
-  id?: string;
-  name?: string;
-  ariaLabel?: string;
-  className?: string;
-  value?: string;
-  defaultValue?: string;
-  disabled?: boolean;
-  onChange?: (value: string) => void;
-  onKeyDown?: (event: React.KeyboardEvent<HTMLInputElement>) => void;
-  onPaste?: (event: React.ClipboardEvent<HTMLInputElement>) => void;
-
-}
