+++ src/types/index.ts
@@ -1,2 +1,4 @@
-export * from './forms';
-export * from './useOtpInput';
+export * from './components.types';
+export * from './useObtainOtp.types';
+export * from './useOtpInput.types';
+export * from './useOtpForm.types';
