+++ src/types/useObtainOtp.types.ts
@@ -1,0 +1,5 @@
+export interface UseObtainOtpReturn {
+  otp: string | null;
+  isError: boolean;
+  errorMessage: string | null;
+}
